Flat IDT + electrode + WG + taper + alignment marker sweep
Version 1 of a GDSFactory file that automatically creates RF pads, flat IDT's, waveguide and waveguide taper (linear). It then places an etch mask to allow protection of the electrode structures during etching of lower layers. FInally four alignment markers are added. This is just the first rough version of this.